AS3 into a rust library
