+++ rust/Cargo.toml
@@ -1,11 +1,19 @@
 [package]
-name = "appcovecompiler"
+name = "as3"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "as3"
+
+[[bin]]
+name = "as3"
+path = "src/main.rs"
+
 [dependencies]
+clap = { version = "4.0.29", features = ["derive"] }
 rayon = "1.6.0"
 regex = "1.7.0"
 serde = {version= "1.0.147", features=["derive"]}
